requirements.txt
Flask-Celery-py3
celery
celery-flower
click
mapchete>=0.21
pysqlite3